requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
fake-useragent==2.0.3
fastapi==0.100.0
frozenlist==1.5.0
gotrue==2.11.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hypercorn==0.14.4
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-postgres==0.0.13
langchain-text-splitters==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.8
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.63.2
orjson==3.10.15
packaging==24.2
pgvector==0.3.6
postgrest==0.19.3
priority==2.0.0
propcache==0.2.1
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.27.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
